requirements.txt
asgiref==3.7.2
decouple==0.0.7
Django==4.2.5
django-appconf==1.0.5
django-imagekit==5.0.0
mysqlclient==2.2.0
pilkit==3.0
Pillow==10.1.0
sqlparse==0.4.4
